perf: remove redis cache lookup
We're doing this lookup on _every event_. It's not slow but on aggregate it ends up being a lot of time.

Looking at production data, it seems $set_once is defined on about 80% of events, so we'll being hitting PG here anyway most of the time. It's probably worth just removing this optimization anyway.

The comment in the code that has been removed about out of order processing is only relevant if the user is e.g. backfilling. This feels like an edge case that isn't top prio atm. Real time events ingested in order will be processed in order, partitioned by distinct_id.